Building object files only instead of libs
In order to avoid having to support a multiplatform build, this
revision just compiles the source. Each platform will have its own
type of build, so it's not worth the effort to address a
multiplatform build, but we still want some testing of the shared
bits, so this facilitates that.