Tweaking bulleted lists, primarily for UITextView
It seems that a UITextView does not accept an NSAttributedString
with a CFParagraphStyle set on it. Internally, it will try to
set headIndent the attributes and fail since it does not have
that method. This is a temporary workaround until I get to the
bottom of the problem with that, since the workaround is not ideal