Making concessions so both renderings look similar
The UITextView was not responding to the styling I added, and
I was having to work around that fact in the BPMarkdownView in ways
that are probably not future-proof. This revision is a compromise
in that it is the intersection of functionality that both views
support.

This basically adds nested lists that are indented correctly and
whose bullets are colored correctly.