Prevent auto-adding trailing whitespace to document.
Prior to this change the Android layout formating code (in Bypass.java)
appended a trailing double-newline to the last element in the formatted
document. This had the implication of affecting the client side UI, as
the target textbox would expand at the bottom by the height of an empty
line.

This change prevents the layout formatter from appending said newlines
by (recursively) keeping track of the index of the currently formatted
element, the number of siblings in the batch and their parent.